Chore: UI ci setup (#4)
This pull request introduces significant changes to the CI workflows for
the API and UI components, including the addition of a new path change
detector workflow. The most important changes include modifying the
concurrency group naming convention, adding a path change detector job,
and setting up a new CI pipeline for the UI.

Modifications to CI workflows:

*
[`.github/workflows/api-ci.yml`](diffhunk://#diff-4d5b51f0d2cf831dd4f3c8083cc1eb5bc1c7e0c1a3d94311d6705579d70cef10L13-R29):
Updated the concurrency group naming to include "api-" and added a
`path-check` job to detect changes in the `app/api` path. The
`Build_and_Test` job now depends on the `path-check` job and only runs
if changes are detected.

*
[`.github/workflows/path-change-detector.yml`](diffhunk://#diff-8750793a0a797b27e3057d69eae99f55d6877bee92648c7cdba109d2fe4e2ae4R1-R47):
Introduced a new workflow to detect changes in specified paths. This
workflow is called by other workflows to determine if specific paths
have been modified.

*
[`.github/workflows/ui-ci.yml`](diffhunk://#diff-d92b8914d50115a215962428fbf419141622feb19bb1b9103d44f644c32dc68eR1-R74):
Created a new CI pipeline for the UI, including a `path-check` job to
detect changes in the `app/ui` path. The pipeline includes steps for
checking out the repository, setting up Flutter, caching dependencies,
building the project, running tests, and performing a SonarQube scan.