feat: Implement HomePage with product display and navigation features (#3)
This pull request includes significant changes to the CI pipelines and
the implementation of a product fetching feature in the UI. The most
important changes include adding a path change detector workflow,
updating the API and UI CI pipelines, and implementing the product
fetching functionality in the UI.

### CI Pipeline Improvements:
*
[`.github/workflows/api-ci.yml`](diffhunk://#diff-4d5b51f0d2cf831dd4f3c8083cc1eb5bc1c7e0c1a3d94311d6705579d70cef10L13-R29):
Added a `path-check` job to detect changes in the `app/api` path and
conditionally run the `Build_and_Test` job based on detected changes.
*
[`.github/workflows/path-change-detector.yml`](diffhunk://#diff-8750793a0a797b27e3057d69eae99f55d6877bee92648c7cdba109d2fe4e2ae4R1-R47):
Introduced a new workflow for detecting changes in specified paths,
which outputs whether changes were detected.
*
[`.github/workflows/ui-ci.yml`](diffhunk://#diff-d92b8914d50115a215962428fbf419141622feb19bb1b9103d44f644c32dc68eR1-R76):
Added a new UI CI pipeline with a `path-check` job to detect changes in
the `app/ui` path and conditionally run the `sonarqube` job for code
analysis.

### Product Fetching Feature:
*
[`app/ui/lib/data/datasources/remote/product_remote_data_source.dart`](diffhunk://#diff-b56bbe7d6b1e2c2f5979a80cac5be207207a5667c59bbe0eea77f013a563c05fR1-R25):
Implemented `ProductRemoteDataSource` to fetch products from a remote
API.
*
[`app/ui/lib/data/models/product_item_model.dart`](diffhunk://#diff-74bd529bc4b0e66ebb359e60e1d2eaa0723efed1d4d54afaa4a18fe2ae2f2dc9R1-R43):
Added `ProductItemModel` to represent product data and convert it to
`ProductItemEntity`.
*
[`app/ui/lib/data/repositories/product_repository_impl.dart`](diffhunk://#diff-a04b2d5c76f107955ee48a61b56535d8b8ac2e73b26166c2cbdf99311fee467cR1-R23):
Implemented `ProductRepositoryImpl` to interact with
`ProductRemoteDataSource` and return product entities.
*
[`app/ui/lib/domain/usecases/get_products_use_case.dart`](diffhunk://#diff-5d05488712a0d6290560188ef63d7f8f3322295e27e71a597c4bbe5f5b57c1c2R1-R16):
Added `GetProductsUseCase` to encapsulate the logic for fetching
products.
*
[`app/ui/lib/presentation/cubits/products/products_cubit.dart`](diffhunk://#diff-7743da84e409514aa43723633908068bd34d6da8a3431816e40b52c1938ead81R1-R66):
Created `ProductsCubit` to manage the state of product fetching and
handle pagination.