Feat(ui): update bottom navigation bar (#8)
This pull request includes changes to the
`app_bottom_navigation_bar.dart` and its corresponding test file to
update the icons and labels in the bottom navigation bar. The most
important changes include replacing the 'Categories' icon with a
'Profile' icon and the 'Profile' icon with a 'Menu' icon.

Updates to bottom navigation bar:

*
[`app/ui/lib/presentation/widgets/app_bottom_navigation_bar.dart`](diffhunk://#diff-d678d79cc1dcbd172c634dfdb4c8c0464f7aead3007618410ec91a5d3f3df393L18-R27):
Replaced the 'Categories' icon and label with 'Profile', and the
'Profile' icon and label with 'Menu'.

Updates to tests:

*
[`app/ui/test/presentation/widgets/app_bottom_navigation_bar_test.dart`](diffhunk://#diff-bfa7701439072f01d07d19303b9b28c1cc42c4143c5ce69d342865fb4ac58e92L27-R27):
Updated the test cases to reflect the changes in the bottom navigation
bar by replacing 'Categories' with 'Menu'.
[[1]](diffhunk://#diff-bfa7701439072f01d07d19303b9b28c1cc42c4143c5ce69d342865fb4ac58e92L27-R27)
[[2]](diffhunk://#diff-bfa7701439072f01d07d19303b9b28c1cc42c4143c5ce69d342865fb4ac58e92L53-R53)