Fix(api): retrieve search products (#15)
This pull request includes changes to the `ProductRepositoryAdapter` and
`DataLoader` classes to improve the handling and initialization of
products in the database. The most important changes include indexing
saved products and removing unnecessary code in the `DataLoader` class.

Improvements to product handling:

*
[`app/api/src/main/java/com/somba/api/adapter/repositories/ProductRepositoryAdapter.java`](diffhunk://#diff-9ac5503f541a97b8a354a539669c6fb15c900fa4e3d5ea44aa5bfb7e9ede4db0L43-R51):
Modified the `save` method to index the saved product by calling the
`index` method after saving.

Code cleanup:

*
[`app/api/src/main/java/com/somba/api/infrastructure/config/DataLoader.java`](diffhunk://#diff-5b49ab7482f0a5b029fdee60efc3bc836e0803b08d7642cc8946bee0e5491cdaL24):
Removed the creation and saving of a separate `products` list in the
`initDatabase` method, simplifying the initialization process.
[[1]](diffhunk://#diff-5b49ab7482f0a5b029fdee60efc3bc836e0803b08d7642cc8946bee0e5491cdaL24)
[[2]](diffhunk://#diff-5b49ab7482f0a5b029fdee60efc3bc836e0803b08d7642cc8946bee0e5491cdaL39)